Factor out source preparation tests
