+++ tests/Functional/Application/PrepareSourceTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Functional\Controller;
+namespace App\Tests\Functional\Application;
 
 use App\Entity\FileSource;
 use App\Entity\GitSource;
@@ -10,66 +10,28 @@ use App\Entity\RunSource;
 use App\Enum\RunSource\State;
 use App\Enum\Source\Type;
 use App\Repository\RunSourceRepository;
-use App\Services\RunSourceSerializer;
-use App\Services\Source\Store;
+use App\Tests\Application\AbstractPrepareSourceTest;
 use App\Tests\DataProvider\TestConstants;
-use App\Tests\DataProvider\UpdateSourceInvalidRequestDataProviderTrait;
-use App\Tests\DataProvider\UpdateSourceSuccessDataProviderTrait;
-use App\Tests\Services\EntityRemover;
-use App\Tests\Services\FileStoreFixtureCreator;
-use League\Flysystem\FilesystemOperator;
+use App\Tests\Services\SourceUserIdMutator;
 
-class UserSourceControllerTest extends AbstractSourceControllerTest
+class PrepareSourceTest extends AbstractPrepareSourceTest
 {
-    use UpdateSourceInvalidRequestDataProviderTrait;
-    use UpdateSourceSuccessDataProviderTrait;
+    use GetClientAdapterTrait;
 
+    private SourceUserIdMutator $sourceUserIdMutator;
     private RunSourceRepository $runSourceRepository;
-    private Store $store;
-    private FileStoreFixtureCreator $fixtureCreator;
-    private FilesystemOperator $runSourceStorage;
-    private FilesystemOperator $fixtureStorage;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $store = self::getContainer()->get(Store::class);
-        \assert($store instanceof Store);
-        $this->store = $store;
+        $sourceUserIdMutator = self::getContainer()->get(SourceUserIdMutator::class);
+        \assert($sourceUserIdMutator instanceof SourceUserIdMutator);
+        $this->sourceUserIdMutator = $sourceUserIdMutator;
 
         $runSourceRepository = self::getContainer()->get(RunSourceRepository::class);
         \assert($runSourceRepository instanceof RunSourceRepository);
         $this->runSourceRepository = $runSourceRepository;
-
-        $fixtureCreator = self::getContainer()->get(FileStoreFixtureCreator::class);
-        \assert($fixtureCreator instanceof FileStoreFixtureCreator);
-        $this->fixtureCreator = $fixtureCreator;
-
-        $runSourceStorage = self::getContainer()->get('run_source.storage');
-        \assert($runSourceStorage instanceof FilesystemOperator);
-        $this->runSourceStorage = $runSourceStorage;
-
-        $fixtureStorage = self::getContainer()->get('test_fixtures.storage');
-        \assert($fixtureStorage instanceof FilesystemOperator);
-        $this->fixtureStorage = $fixtureStorage;
-
-        $entityRemover = self::getContainer()->get(EntityRemover::class);
-        if ($entityRemover instanceof EntityRemover) {
-            $entityRemover->removeAll();
-        }
-    }
-
-    public function testPrepareRunSource(): void
-    {
-        $fileSource = new FileSource($this->authenticationConfiguration->authenticatedUserId, 'file source label');
-        $source = new RunSource($fileSource);
-
-        $this->store->add($source);
-
-        $response = $this->applicationClient->makePrepareSourceRequest($this->validToken, $source->getId(), []);
-
-        $this->responseAsserter->assertNotFoundResponse($response);
     }
 
     /**
@@ -86,7 +48,11 @@ class UserSourceControllerTest extends AbstractSourceControllerTest
         $this->sourceUserIdMutator->setSourceUserId($source);
         $this->store->add($source);
 
-        $response = $this->applicationClient->makePrepareSourceRequest($this->validToken, $source->getId(), $payload);
+        $response = $this->applicationClient->makePrepareSourceRequest(
+            $this->authenticationConfiguration->validToken,
+            $source->getId(),
+            $payload
+        );
 
         $runSource = $this->runSourceRepository->findByParent($source);
         self::assertInstanceOf(RunSource::class, $runSource);
@@ -167,27 +133,5 @@ class UserSourceControllerTest extends AbstractSourceControllerTest
                 ],
             ],
         ];
-    }
-
-    public function testReadSuccess(): void
-    {
-        $serializedRunSourceFixturePath = 'RunSource/source_yml_yaml_entire.yaml';
-
-        $fileSource = new FileSource($this->authenticationConfiguration->authenticatedUserId, 'file source label');
-        $runSource = new RunSource($fileSource);
-        $this->store->add($runSource);
-
-        $this->fixtureCreator->copyTo(
-            $serializedRunSourceFixturePath,
-            $this->runSourceStorage,
-            $runSource->getDirectoryPath() . '/' . RunSourceSerializer::SERIALIZED_FILENAME
-        );
-
-        $response = $this->applicationClient->makeReadSourceRequest($this->validToken, $runSource->getId());
-
-        $this->responseAsserter->assertReadSourceSuccessResponse(
-            $response,
-            trim($this->fixtureStorage->read($serializedRunSourceFixturePath))
-        );
     }
 }
